# openapi.yaml
openapi: 3.1.0
info:
  title: Trustify
  description: Software Supply-Chain Security API
  license:
    name: Apache License, Version 2.0
    identifier: Apache-2.0
  version: 0.6.0-rc.1
paths:
...
  /api/v3/user:
    get:
      tags:
      - tenancy
      summary: List users
      operationId: listUsers
      parameters:
      - name: q
        in: query
        description: |
          EBNF grammar for the _q_ parameter:
          ```text
              q = ( values | filter ) { '&' q }
              values = value { '|', values }
              filter = field, operator, values
              operator = "=" | "!=" | "~" | "!~" | ">=" | ">" | "<=" | "<"
              value = (* any text but escape special characters with '\' *)
              field = (* must match an entity attribute name *)
          ```
          Any values in a _q_ will result in a case-insensitive "full
          text search", effectively producing an OR clause of LIKE
          clauses for every string-ish field in the resource being
          queried.

          Examples:
          - `foo` - any field containing 'foo'
          - `foo|bar` - any field containing either 'foo' OR 'bar'
          - `foo&bar` - some field contains 'foo' AND some field contains 'bar'

          A _filter_ may also be used to constrain the results. The
          filter's field name must correspond to one of the resource's
          attributes. If it doesn't, an error will be returned
          containing a list of the valid fields for that resource.

          An ASCII value of `NUL`, percent-encoded as `%00`, may be used
          to find resources on which a particular field isn't set. For
          example, `name=%00` and `name!=%00` yield the WHERE clauses,
          'NAME IS NULL' and 'NAME IS NOT NULL', respectively.

          Examples:
          - `name=foo` - entity's _name_ matches 'foo' exactly
          - `name~foo` - entity's _name_ contains 'foo', case-insensitive
          - `name~foo|bar` - entity's _name_ contains either 'foo' OR 'bar', case-insensitive
          - `name=` - entity's _name_ is the empty string, ''
          - `name=%00` - entity's _name_ isn't set
          - `published>3 days ago` - date values can be "human time"

          Multiple full text searches and/or filters should be
          '&'-delimited -- they are logically AND'd together.

          - `red hat|fedora&labels:type=cve|osv&published>last wednesday 17:00`

          Fields corresponding to JSON objects in the database may use a
          ':' to delimit the column name and the object key,
          e.g. `purl:qualifiers:type=pom`

          Any operator or special character, e.g. '|', '&', within a
          value should be escaped by prefixing it with a backslash.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: |
          EBNF grammar for the _sort_ parameter:
          ```text
              sort = field [ ':', order ] { ',' sort }
              order = ( "asc" | "desc" )
              field = (* must match the name of entity's attributes *)
          ```
          The optional _order_ should be one of "asc" or "desc". If
          omitted, the order defaults to "asc".

          Each _field_ name must correspond to one of the columns of the
          table holding the entities being queried. Those corresponding
          to JSON objects in the database may use a ':' to delimit the
          column name and the object key,
          e.g. `purl:qualifiers:type:desc`
        required: false
        schema:
          type: string
      - name: offset
        in: query
        description: |-
          The first item to return, skipping all that come before it.

          NOTE: The order of items is defined by the API being called.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        description: |-
          The maximum number of entries to return.

          Zero means: return no items (the total count is still computed if requested).
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: total
        in: query
        description: Whether to compute and return the total count of matching items.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Matching users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResults_User'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
  /api/v3/user/{email}:
    get:
      tags:
      - tenancy
      summary: Get a user
      operationId: getUser
      parameters:
      - name: email
        in: path
        description: The e-mail address of the user
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The user
          headers:
            etag:
              schema:
                type: string
              description: Revision ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '404':
          description: The user was not found
    put:
      tags:
      - tenancy
      summary: Create or update a user
      operationId: putUser
      parameters:
      - name: email
        in: path
        description: The e-mail address of the user
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision to update
        required: false
        schema:
          type:
          - string
          - 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '200':
          description: The user was updated
          headers:
            etag:
              schema:
                type: string
              description: Revision ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '201':
          description: The user was created
          headers:
            etag:
              schema:
                type: string
              description: Revision ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '409':
          description: The external ID is already in use
        '412':
          description: The provided If-Match revision did not match the actual revision
    delete:
      tags:
      - tenancy
      summary: Delete a user, including its team memberships and role bindings
      operationId: deleteUser
      parameters:
      - name: email
        in: path
        description: The e-mail address of the user
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision to delete
        required: false
        schema:
          type:
          - string
          - 'null'
      responses:
        '204':
          description: The user was deleted, or did not exist
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '412':
          description: The provided If-Match revision did not match the actual revision
  /api/v3/user/{email}/access:
    get:
      tags:
      - tenancy
      summary: Get the role bindings applying to a user, directly or through teams
      operationId: getUserAccess
      parameters:
      - name: email
        in: path
        description: The e-mail address of the user
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The role bindings applying to the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Access'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '404':
          description: The user was not found
  /api/v3/user/{email}/change-email:
    post:
      tags:
      - tenancy
      summary: Change the e-mail address of a user
      operationId: changeUserEmail
      parameters:
      - name: email
        in: path
        description: The current e-mail address of the user
        required: true
        schema:
          type: string
      - name: if-match
        in: header
        description: The revision to update
        required: false
        schema:
          type:
          - string
          - 'null'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeEmailRequest'
        required: true
      responses:
        '200':
          description: The e-mail address was changed
          headers:
            etag:
              schema:
                type: string
              description: Revision ID
            location:
              schema:
                type: string
              description: The new location of the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: The request was not valid
        '401':
          description: The user was not authenticated
        '403':
          description: The user authenticated, but not authorized for this operation
        '404':
          description: The user was not found
        '409':
          description: The new e-mail address is already in use
        '412':
          description: The provided If-Match revision did not match the actual revision
...
components:
  schemas:
    Access:
      type: object
      description: A role a user holds on a group, and all groups below it.
      required:
      - group
      - role
      - via
      properties:
        group:
          type: string
          description: The ID of the group the role was granted on.
        role:
          $ref: '#/components/schemas/Role'
        via:
          $ref: '#/components/schemas/Via'
...
    ChangeEmailRequest:
      type: object
      description: Request to change the e-mail address of a user.
      required:
      - newEmail
      properties:
        newEmail:
          type: string
...
    PaginatedResults_User:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            type: object
            description: A user, identified by their e-mail address.
            required:
            - id
            - email
            - state
            - createdAt
            properties:
              createdAt:
                type: string
                format: date-time
              displayName:
                type:
                - string
                - 'null'
              email:
                type: string
                description: The e-mail address, lower-cased.
              externalId:
                type:
                - string
                - 'null'
                description: An optional ID, assigned by an external system.
              id:
                type: string
                description: The ID of the user, assigned by the system.
              lastLogin:
                type:
                - string
                - 'null'
                format: date-time
              state:
                $ref: '#/components/schemas/UserState'
        total:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
...
    Role:
      type: string
      description: A role on an SBOM group, each role includes all permissions of the ones before it.
      enum:
      - viewer
      - uploader
      - editor
      - admin
...
    User:
      type: object
      description: A user, identified by their e-mail address.
      required:
      - id
      - email
      - state
      - createdAt
      properties:
        createdAt:
          type: string
          format: date-time
        displayName:
          type:
          - string
          - 'null'
        email:
          type: string
          description: The e-mail address, lower-cased.
        externalId:
          type:
          - string
          - 'null'
          description: An optional ID, assigned by an external system.
        id:
          type: string
          description: The ID of the user, assigned by the system.
        lastLogin:
          type:
          - string
          - 'null'
          format: date-time
        state:
          $ref: '#/components/schemas/UserState'
    UserRequest:
      type: object
      description: Mutable properties of a [`User`].
      properties:
        disabled:
          type: boolean
          description: Prevent the user from signing in. Existing role bindings are kept.
        displayName:
          type:
          - string
          - 'null'
        externalId:
          type:
          - string
          - 'null'
          description: An optional ID, assigned by an external system, unique across all users.
    UserState:
      type: string
      enum:
      - invited
      - active
      - disabled
...
    Via:
      oneOf:
      - type: object
        description: Directly bound to the user.
        required:
        - kind
        properties:
          kind:
            type: string
            enum:
            - direct
      - type: object
        description: Bound to a team the user is a member of.
        required:
        - id
        - kind
        properties:
          id:
            type: string
            description: The ID of the team
          kind:
            type: string
            enum:
            - team
      description: How access was granted to a user.
...